Store epoch day in 32-bit integer.
Because it can maximum date (31.12.9999) can fit in 32 bits (max epoch day = 2932896).

+++ shared/src/main/java/com/pelmenstar/projktSens/shared/time/PrettyDateFormatter.java
@@ -50,9 +50,9 @@ public abstract class PrettyDateFormatter {
      */
     public final void appendPrettyDate(@ShortDateInt int date, @NotNull StringBuilder sb) {
         boolean appendDate = true;
-        long nowEpochDay = ShortDate.nowEpochDay();
+        int nowEpochDay = ShortDate.nowEpochDay();
 
-        switch ((int)(ShortDate.toEpochDay(date) - nowEpochDay)) {
+        switch (ShortDate.toEpochDay(date) - nowEpochDay) {
             case 0: {
                 sb.append(getTodayString());
                 appendDate = false;
